Empty i18n __init__.py
