+++ kalite/facility/views.py
@@ -18,7 +18,7 @@ from .forms import FacilityUserForm, FacilityGroupForm
 from .models import Facility, FacilityGroup, FacilityUser
 from fle_utils.internet.functions import set_query_params
 from kalite.dynamic_assets.decorators import dynamic_settings
-from kalite.i18n import get_default_language
+from kalite.i18n.base import get_default_language
 from kalite.shared.decorators.auth import require_authorized_admin
