+++ kalite/i18n/management/commands/create_dummy_language_pack.py
@@ -20,7 +20,7 @@ from django.conf import settings
 from django.core.management.base import NoArgsCommand
 
 from fle_utils.general import ensure_dir
-from kalite.i18n import get_language_pack_url, get_locale_path, update_jsi18n_file
+from kalite.i18n.base import get_language_pack_url, get_locale_path, update_jsi18n_file
 from kalite.version import VERSION
 
 logging = settings.LOG
