+++ kalite/i18n/tests/create_dummy_language_pack_tests.py
@@ -4,7 +4,7 @@ import zipfile
 from cStringIO import StringIO
 from mock import patch, MagicMock, call, mock_open
 
-from kalite.i18n import get_language_pack_url
+from kalite.i18n.base import get_language_pack_url
 from kalite.i18n.management.commands import create_dummy_language_pack as mod
 from kalite.testing.base import KALiteTestCase
