+++ kalite/updates/api_views.py
@@ -26,7 +26,7 @@ from fle_utils.general import isnumeric, break_into_chunks, softload_json
 from fle_utils.internet.decorators import api_handle_error_with_json
 from fle_utils.internet.classes import JsonResponse, JsonResponseMessageError, JsonResponseMessageSuccess
 from fle_utils.orderedset import OrderedSet
-from kalite.i18n import get_youtube_id, lcode_to_ietf, delete_language, get_language_name
+from kalite.i18n.base import get_youtube_id, lcode_to_ietf, delete_language, get_language_name
 from kalite.shared.decorators.auth import require_admin
 from kalite.topic_tools.settings import TOPICS_FILEPATHS, CHANNEL
 from kalite.topic_tools.content_models import get_topic_update_nodes
