+++ kalite/updates/management/commands/languagepackdownload.py
@@ -20,7 +20,7 @@ from fle_utils.internet.download import callback_percent_proxy, download_file
 from django.core.management import call_command
 from django.core.management.base import CommandError
 from django.utils.translation import ugettext as _
-from kalite.i18n import get_language_name, get_language_pack_url, \
+from kalite.i18n.base import get_language_name, get_language_pack_url, \
     get_localized_exercise_dirpath, get_po_filepath, get_srt_path, \
     lcode_to_django_dir, lcode_to_ietf, update_jsi18n_file
 from kalite.version import SHORTVERSION
